requirements.txt
python-telegram-bot==11
python-graphql-client